Chore: Move VerificationPage into feature folder
