+++ lib/main.dart
@@ -3,9 +3,9 @@ import 'package:decentproof/features/hashing/pages/AudioPage.dart';
 import 'package:decentproof/pages/homepage/HomePage.dart';
 import 'package:decentproof/features/intro/pages/IntroPage.dart';
 import 'package:decentproof/features/intro/logic/firstTimeUserCheck.dart';
-import 'package:decentproof/pages/settingspage/SettingsPage.dart';
+import 'package:decentproof/features/settings/pages/SettingsPage.dart';
 import 'package:decentproof/pages/submissionpage/SubmissionPage.dart';
-import 'package:decentproof/pages/verificationpage/VerificationPage.dart';
+import 'package:decentproof/features/verification/pages/VerificationPage.dart';
 import 'package:decentproof/pages/videoimagepage/VideoImagePage.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:firebase_app_check/firebase_app_check.dart';
